added auth

+++ base/views.py
@@ -3,6 +3,35 @@ from django.http import HttpResponse
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib import messages
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+
+def loginPage(request):
+    if(request.method == 'POST'):
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user =  User.objects.get(username=username)
+        except:
+            messages.error(request, "username not found")
+
+        user = authenticate(request, username=username, password=password) 
+
+        if user is not None:
+            login(request,user)   
+            return redirect('home')
+        else:
+            messages.error(request, "username or password doesnt exists")
+
+    return render(request, 'base/login_register.html')
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ""
@@ -15,6 +44,7 @@ def room(request,pk):
     room = Room.objects.get(id=pk)
     return render(request,'base/room.html', {'room' : room})
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -24,9 +54,14 @@ def createRoom(request):
             return redirect('home')
     return render(request, 'base/room-form.html', {'form': form})
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to change')
+
     
     if request.method == 'POST':
        form = RoomForm(request.POST, instance=room)
@@ -36,9 +71,13 @@ def updateRoom(request, pk):
        
     return render(request, 'base/room-form.html', {'form': form})
 
+@login_required(login_url='/login')
 def deletRoom(request, pk):
     room = Room.objects.get(id=pk)
     
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to change')
+
     if request.method == 'POST':
         room.delete()
         return redirect("home")
